Make the navigation smoother
using the navigation at the bottom of the page: if you are inside a section, the right arrow link will jump to the next section. If you are inside a subsection, the right arrow link will jump to the next sub section. Small wording fixes are also done in this commits